replace the manifest VersionProvider with the properties VersionProvider
